cursor-ai-devops/backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
websockets>=11.0.0
langchain>=0.1.0
chromadb>=0.4.0
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0
kubernetes>=27.2.0
docker>=6.1.0
gitpython>=3.1.0
prometheus-client>=0.17.0
sqlmodel>=0.0.14
aiosqlite>=0.19.0
watchdog>=3.0.0
pydantic>=2.5.0
openai>=1.3.0
anthropic>=0.7.0
ollama>=0.1.0
httpx>=0.25.0
python-multipart>=0.0.6
PyYAML>=6.0.1
Jinja2>=3.1.2
terraform-compliance>=1.3.0
pytest>=7.4.0
pytest-asyncio>=0.21.1

